TO-REVERT: debug missing core config parsing
This patch instruments the code such that we can identify where the
code.* settings are not parsed. The purpose: To make reasonably sure
that we caught at least the vast majority of code paths in the previous
commit (where we tried to add missing calls to ensure in particular that
core.longPaths was heeded by all command) by running the test suite.

Of course, this commit only adds debugging code which we will presently
revert, so that it 1) does not clutter Git for Windows' source code, yet
2) is still in the commit history.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>